with simple performance test https://gist.github.com/ishidawataru/fe79cd690a4d97030ae6

before 
./inject  0.20s user 0.12s system 1% cpu 19.146 total

after 
./inject  0.25s user 0.09s system 44% cpu 0.788 total